refactor(desktop): move leaf domain components into features/ dirs
budget, notifications, skills, plans, phases, permissions components
moved to features/<domain>/components/<Name>/index.tsx.
