+++ apps/desktop/src/components/WorkspaceSettingsDialog.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import type { WorkspaceId } from '@kay-am/types';
 import { Button, Dialog, Input, cn } from '@kay-am/ui';
 import { FolderCode, GitBranch, Unplug, Zap } from 'lucide-react';
-import { SkillsPanel } from './SkillsPanel';
-import { PhasesPanel } from './PhasesPanel';
+import { SkillsPanel } from '../features/skills/components/SkillsPanel';
+import { PhasesPanel } from '../features/phases/components/PhasesPanel';
 import { BulkSessionDeleteDialog } from './WorkspacesSidebar';
 import { formatError } from '../shared/lib/errors';
 import { DEFAULT_BRANCH_PREFIX, settingBranchPrefix } from '../features/settings/settings';
