+++ apps/desktop/src/features/budget/components/BudgetRulesPanel/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { Button, Input, Select } from '@kay-am/ui';
 import type { BudgetRule, ProviderName } from '@kay-am/types';
-import { formatError } from '../shared/lib/errors';
-import { useAppStore } from '../store';
+import { formatError } from '../../../../shared/lib/errors';
+import { useAppStore } from '../../../../store';
 
 const PROVIDER_OPTIONS: ProviderName[] = ['anthropic', 'cursor', 'codex'];
