+++ apps/desktop/src/features/notifications/components/NotificationCenter/index.tsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom';
 import { Bell, CheckCircle, AlertCircle, AlertTriangle, Info, X, Trash2 } from 'lucide-react';
 import { Tooltip, cn } from '@kay-am/ui';
 import type { Notification, NotificationSeverity } from '@kay-am/db';
-import { useAppStore } from '../store';
+import { useAppStore } from '../../../../store';
 
 function severityIcon(severity: NotificationSeverity, size = 13) {
   switch (severity) {
