+++ apps/desktop/src/features/permissions/components/DiffViewSelector/index.tsx
