+++ apps/desktop/src/features/phases/components/PhasesPanel/index.tsx
@@ -2,9 +2,12 @@ import { useEffect, useState } from 'react';
 import { Button, Input, Select, Textarea } from '@kay-am/ui';
 import type { Step, StepId, Workflow, WorkflowId, WorkspaceId } from '@kay-am/types';
 import type { ProviderId } from '@kay-am/types';
-import { formatError } from '../shared/lib/errors';
-import { EMPTY_ARRAY, useAppStore } from '../store';
-import type { PhaseTemplateUpsertArgs, PhaseDefinitionUpsertArgs } from '../features/phases/phases';
+import { formatError } from '../../../../shared/lib/errors';
+import { EMPTY_ARRAY, useAppStore } from '../../../../store';
+import type {
+  PhaseTemplateUpsertArgs,
+  PhaseDefinitionUpsertArgs,
+} from '../../../../features/phases/phases';
 
 interface PhasesPanelProps {
   readonly workspaceId: WorkspaceId;
