+++ apps/desktop/src/features/skills/components/SkillsPanel/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { Button, Input, Textarea } from '@kay-am/ui';
 import type { Skill, SkillFrontmatter, WorkspaceId } from '@kay-am/types';
-import { formatError } from '../shared/lib/errors';
-import { EMPTY_ARRAY, useAppStore } from '../store';
+import { formatError } from '../../../../shared/lib/errors';
+import { EMPTY_ARRAY, useAppStore } from '../../../../store';
 
 interface SkillsPanelProps {
   readonly workspaceId: WorkspaceId;
